requirement.txt
blinker==1.6.2
click==8.1.7
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
psycopg2-binary==2.9.7
PyJWT==2.8.0
SQLAlchemy==2.0.21
typing_extensions==4.8.0
Werkzeug==2.3.7
